Clean up extension functions and documentation
