+++ library/encoding-core/src/commonMain/kotlin/io/matthewnelson/encoding/core/Decoder.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-@file:Suppress("MemberVisibilityCanBePrivate")
+@file:Suppress("SpellCheckingInspection")
 
 package io.matthewnelson.encoding.core
