+++ library/encoding-core/src/commonMain/kotlin/io/matthewnelson/encoding/core/ExperimentalEncodingApi.kt
@@ -13,6 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
+@file:Suppress("SpellCheckingInspection")
+
 package io.matthewnelson.encoding.core
 
 /**
